requirement.txt
fastdtw>=0.3.2
numpy>=1.14.3
pandas>=0.23.0
Python>=3.6.5
scikit-learn>=0.19.1
scipy>=1.1.0
seaborn>=0.9.0
tensorflow==1.9.0
Keras==2.0.8
imbalanced-learn>=0.3.3
lightgbm==2.0.10
paramiko==2.4.1
ffmpeg-python>=0.1.16
